output box changes, css centering and more
moved output box to the top left as an absolute element, fixed
package.json to avoid overlapping of the box with the champ table.

changed how the output box's <br> were handled (now it's consistent and
the info display is more compact).

moved most of inline styles to css, changed some css properties.
centered everything thru css.

as of now the script for importing the sets won't work.